requirements.txt
numpy
Pillow

torch
torchvision

fastapi
uvicorn
typing-extensions
python-multipart